extract all category Tests in own package and different files
